Add support for reboot in interactive mode
The output in interactive mode is forwarded to tmt's output rather than
to a log, hence the previous approach of detecting reboot based on
specific string in the test output was not sufficient. Switch to a
different mechanism to overcome this -- a new file signalling a reboot
request is created by the test which the executor can detect and act
upon after the file is pulled from the guest.

+++ tests/execute/reboot/basic.sh
@@ -8,22 +8,27 @@ rlJournalStart
         rlRun "pushd data"
     rlPhaseEnd
 
-    rlPhaseStartTest "Simple reboot test"
-        rlRun -s "tmt run -i $run -dddvvv"
-        rlAssertGrep "Reboot during test '/test' with reboot count 1" $rlRun_LOG
-        rlAssertGrep "After first reboot" $rlRun_LOG
-        rlAssertGrep "Reboot during test '/test' with reboot count 2" $rlRun_LOG
-        rlAssertGrep "After second reboot" $rlRun_LOG
-        rlAssertGrep "Reboot during test '/test' with reboot count 3" $rlRun_LOG
-        rlAssertGrep "After third reboot" $rlRun_LOG
-        rlRun "rm $rlRun_LOG"
+    for interactive in "" "--interactive"; do
+        rlPhaseStartTest "Simple reboot test (interactivity: $interactive)"
+            rlRun -s "tmt run --scratch -i $run -dddvvva execute -h tmt $interactive"
+            rlAssertGrep "Reboot during test '/test' with reboot count 1" $rlRun_LOG
+            rlAssertGrep "After first reboot" $rlRun_LOG
+            rlAssertGrep "Reboot during test '/test' with reboot count 2" $rlRun_LOG
+            rlAssertGrep "After second reboot" $rlRun_LOG
+            rlAssertGrep "Reboot during test '/test' with reboot count 3" $rlRun_LOG
+            rlAssertGrep "After third reboot" $rlRun_LOG
+            rlRun "rm $rlRun_LOG"
 
-        # Check that the whole output log is kept
-        rlRun "log=$run/plan/execute/data/test/output.txt"
-        rlAssertGrep "After first reboot" $log
-        rlAssertGrep "After second reboot" $log
-        rlAssertGrep "After third reboot" $log
-    rlPhaseEnd
+            # Check that the whole output log is kept
+            # The test output is not stored in log in interactive mode
+            if [ -z "$interactive" ]; then
+                rlRun "log=$run/plan/execute/data/test/output.txt"
+                rlAssertGrep "After first reboot" $log
+                rlAssertGrep "After second reboot" $log
+                rlAssertGrep "After third reboot" $log
+            fi
+        rlPhaseEnd
+    done
 
     rlPhaseStartCleanup
         rlRun "rm -rf output $run" 0 "Remove run directory"
